+++ protocol/messenger.go
@@ -21,8 +21,6 @@ import (
 	"github.com/davecgh/go-spew/spew"
 	"github.com/golang/protobuf/proto"
 
-	bindata "github.com/status-im/migrate/v4/source/go_bindata"
-
 	"github.com/status-im/status-go/appdatabase"
 	"github.com/status-im/status-go/appmetrics"
 	"github.com/status-im/status-go/connection"
@@ -32,7 +30,6 @@ import (
 	"github.com/status-im/status-go/multiaccounts"
 	"github.com/status-im/status-go/multiaccounts/accounts"
 	"github.com/status-im/status-go/protocol/anonmetrics"
-	"github.com/status-im/status-go/protocol/anonmetrics/migrations"
 	"github.com/status-im/status-go/protocol/audio"
 	"github.com/status-im/status-go/protocol/common"
 	"github.com/status-im/status-go/protocol/communities"
@@ -279,9 +276,7 @@ func NewMessenger(
 	// Initialise anon metrics server
 	var anonMetricsServer *anonmetrics.Server
 	if c.anonMetricsServerConfig != nil && c.anonMetricsServerConfig.Enabled {
-		// Generate anonymous metrics postgres migration resource and pass to Server gen
-		postgresMigration := bindata.Resource(migrations.AssetNames(), migrations.Asset)
-		server, err := anonmetrics.NewServer(c.anonMetricsServerConfig.PostgresURI, postgresMigration)
+		server, err := anonmetrics.NewServer(c.anonMetricsServerConfig.PostgresURI)
 		if err != nil {
 			return nil, errors.Wrap(err, "failed to create anonmetrics.Server")
 		}
